postMessage action to subwindow
